test: move query tests to the query file
